refactor(ts): wc consumer type file
