+++ packages/thirdweb/src/wallets/wallet-connect/consumer/request-handlers/send-raw-transaction.ts
@@ -1,6 +1,6 @@
 import type { Hex } from "../../../../utils/encoding/hex.js";
 import type { Account } from "../../../interfaces/wallet.js";
-import type { WalletConnectRawTransactionRequestParams } from "../../types.js";
+import type { WalletConnectRawTransactionRequestParams } from "../types.js";
 
 /**
  * @internal
